introduce new viewModel for countdownFragment with countdownTimer.

+++ app/src/main/java/com/example/pomodoro/ui/notifications/ActivityRecyclerViewAdapter.java
@@ -101,7 +101,7 @@ public class ActivityRecyclerViewAdapter extends RecyclerView.Adapter<ActivityRe
         public String getText() {
             String title = mItem.getTitle();
             String titleLabel = (title.length()>6)? title.substring(0,5)+"...":title;
-            int totalTime = mItem.getTotalTime();
+            long totalTime = mItem.getTotalTime();
 
             String timeLabel = (totalTime/60/60 > 0)? String.format("%02d:", totalTime/60/60):"";
             timeLabel += String.format("%02d:", (totalTime/60)%60)
